feat(ml): Split LLM generator headers and normalize member initializers
- Move ONNX-specific `TextGenerator` declarations into independent header file.
- Normalize default member initializers to consistent brace form
